% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{bisg}
\alias{bisg}
\title{Performs Bayesian Improved Surname Geocoding: Computes the probability a
person is of a specific racial group, conditioned on surname and geolocation.}
\usage{
bisg(
  voter_file,
  surname_col,
  geo_col,
  census_counts = NULL,
  geography = NULL,
  state = NULL,
  county = NULL,
  year = NULL,
  geo_col_counts = "fips",
  race_cols = c("whi", "bla", "his", "asi", "oth"),
  impute_missing = TRUE,
  verbose = FALSE,
  cache = FALSE
)
}
\arguments{
\item{voter_file}{A tibble containing a list of voters (by row), and a
column that denotes their surname.}

\item{surname_col}{A string denoting which column contains the voter surname.}

\item{geo_col}{A string denoting which column contains the geographic unit
ID.}

\item{census_counts}{A tibble containing counts (divided amongst constituent
groups) per geographic units (rows). If NULL, these counts will be obtained
using the eiCompare helper function and the other parameters.}

\item{geography}{The geographic level at which to obtain Census data. If
obtaining data from the decennial Census, can be up to "block". If
obtaining data from the ACS, can only be up to "block group".}

\item{state}{The state from which to obtain Census data, as a string.}

\item{county}{The county(ies) from which to obtain Census data. If NULL,
data is obtained from all counties in the state.}

\item{year}{The year to obtain Census data from. If 2010, uses decennial
data. Otherwise, uses the 5-year ACS summary data.}

\item{geo_col_counts}{A string denoting the column in the counts tibble that
refers to the geographic unit.}

\item{race_cols}{A list of strings denoting the columns containing racial
groups.}

\item{impute_missing}{A bool denoting whether voter file entries that do not
match at the geographic level should be imputed with either the surname
probabilities, or should be imputed with probabilities calculated at a
broader geographic unit.}

\item{verbose}{A boolean denoting the verbosity.}

\item{cache}{A boolean denoting whether Census data should be cached.}
}
\value{
A tibble with rows denoting voters and columns denoting the
 probability that each voter is of a particular racial group.
}
\description{
Performs Bayesian Improved Surname Geocoding: Computes the probability a
person is of a specific racial group, conditioned on surname and geolocation.
}
