% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{compute_p_g}
\alias{compute_p_g}
\title{Computes the probability a person is located in a specific geographic unit.}
\usage{
compute_p_g(counts, cols = c("whi", "bla", "his", "asi", "oth"))
}
\arguments{
\item{counts}{A tibble containing counts (divided amongst constituent groups)
per geographic units (rows).}

\item{cols}{The columns denoting the constituent groups within each
geographic units.}
}
\value{
A vector containing the probability of selecting a person from each
 geographic unit.
}
\description{
This is a utility function for performing BISG. It operates on a dataframe
obtained from the Census Bureau via the provided eiCompare helper function.
}
