% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{compute_p_g_cond_r}
\alias{compute_p_g_cond_r}
\title{Computes the probability a person is in a specific geographic unit,
conditioned on a racial group.}
\usage{
compute_p_g_cond_r(geo_counts, cols = c("whi", "bla", "his", "asi", "oth"))
}
\arguments{
\item{geo_counts}{A tibble containing counts (divided amongst constituent groups)
per geographic units (rows).}

\item{cols}{The columns denoting the constituent groups within each
geographic units.}
}
\value{
A tibble with dimensions equal to counts, with each entry denoting
 the probability of selecting a person per geographic unit (rows)
 conditioned on racial group (columns).
}
\description{
This is a utility function for performing BISG. It operates on a dataframe
obtained from the Census Bureau via the provided eiCompare helper function.
}
