% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{compute_p_r_cond_g}
\alias{compute_p_r_cond_g}
\title{Computes the probability a person is of a specific racial group, conditioned
on geographic unit.}
\usage{
compute_p_r_cond_g(counts, cols = c("whi", "bla", "his", "asi", "oth"))
}
\arguments{
\item{counts}{A tibble containing counts (divided amongst constituent groups)
per geographic units (rows).}

\item{cols}{The columns denoting the constituent groups within each
geographic units.}
}
\value{
A tibble with dimensions equal to counts, with each entry denoting
 the probability of selecting a person per racial group (columns) conditioned
 on racial group (rows).
}
\description{
This is a utility function for performing BISG. It operates on a dataframe
obtained from the Census Bureau via the provided eiCompare helper function.
}
