% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{compute_p_r_cond_s}
\alias{compute_p_r_cond_s}
\title{Computes the probability a person is of a specific racial group, conditioned
on surname.}
\usage{
compute_p_r_cond_s(voter_file, surname_col)
}
\arguments{
\item{voter_file}{A tibble containing a list of voters (by row), and a
column that denotes their surname.}

\item{surname_col}{A string denoting which column contains the voter surname.}
}
\value{
A tibble with rows denoting voters and columns denoting the
 probability that each voter is of a particular racial group.
}
\description{
This is a utility function for performing BISG. It operates on a voter file,
and utilizes WRU's function for calculating probability of race by surname.
}
