% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{compute_product_err}
\alias{compute_product_err}
\title{Calculate error on the product of the terms in the numerator of the BISG
estimate.}
\usage{
compute_product_err(p_g_r, p_r_s, p_g_r_err, p_r_s_err)
}
\arguments{
\item{p_g_r}{A tibble denoting the probability of a geography, given race.
Each row in the tibble denotes a voter.}

\item{p_r_s}{A tibble denoting the probability of a racial group, given a
surname. Each row in the tibble denotes a voter.}

\item{p_g_r_err}{A tibble denoting the error on the probability of a
geography, conditioned on racial group. Same size as p_g_r.}

\item{p_r_s_err}{A tibble denoting the error on the probability of racial
group, conditioned on race. Same size as p_r_s.}
}
\value{
The error on p_g_r times p_r_s.
}
\description{
This serves as a helper function for computing the BISG error estimate.
}
