% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgm.R
\docType{data}
\name{names}
\alias{names}
\title{Converts a character vector into a dataframe with n+1 columns. The first
n columns contain elements of trigrams, and the last contains the original
character vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
names
}
\arguments{
\item{names}{a character vector of names for conversion to letter-level
trigrams}

\item{n}{the n-gram size to use. Defaults to 3 (trigram)}
}
\description{
Converts a character vector into a dataframe with n+1 columns. The first
n columns contain elements of trigrams, and the last contains the original
character vector
}
\keyword{datasets}
