% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgm.R
\name{race_from_letters}
\alias{race_from_letters}
\title{Get race probabilities for a vector of names based on a language model}
\usage{
race_from_letters(names, model = "trigram")
}
\arguments{
\item{names}{a character vector of names for which to retrieve probabilities}

\item{model}{the type of character-level language model to use. Currently
only trigram}
}
\description{
Get race probabilities for a vector of names based on a language model
}
\examples{
names <- c("alexander", "wesley", "smith", "johnson")
race_from_letters(names)
}
