% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_census_geography}
\alias{swap_census_geography}
\title{Gets identifier for geographic unit at the level above the provided unit.}
\usage{
swap_census_geography(census_geo)
}
\arguments{
\item{census_geo}{A string indicating the geographic level.}
}
\value{
A string indicating the geographic unit one level above the provided
 unit.
}
\description{
This function accept a Census geographic identifier and returns the
identifier one level above the provided one, in terms of spatial coverage.
}
