% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisg.R
\name{compute_p_r_cond_s_g}
\alias{compute_p_r_cond_s_g}
\title{Computes the probability a person is of a specific racial group, conditioned
on surname and geolocation.}
\usage{
compute_p_r_cond_s_g(
  voter_file,
  counts,
  surname_col,
  geo_col,
  race_cols = c("whi", "bla", "his", "asi", "oth"),
  geo_col_counts = "fips",
  p_r_s = NULL
)
}
\arguments{
\item{voter_file}{A tibble containing a list of voters (by row), and a
column that denotes their surname.}

\item{counts}{A tibble containing counts (divided amongst constituent groups)
per geographic units (rows).}

\item{surname_col}{A string denoting which column contains the voter surname.}

\item{geo_col}{A string denoting which column contains the geographic unit
ID.}

\item{race_cols}{A list of strings denoting the columns containing racial
groups.}

\item{geo_col_counts}{A string denoting the column in the counts tibble that
refers to the geographic unit.}

\item{p_r_s}{A dataframe denoting the probability of race conditioned on
surname that matches with the provided voter file. Can accelerate
computation if provided up front.}
}
\value{
A tibble with rows denoting voters and columns denoting the
 probability that each voter is of a particular racial group.
}
\description{
This is a utility function for performing BISG. It operates on a voter file,
and counts obtained from the Census Bureau via eiCompare's helper function.
}
